using double values to show cpu usage

+++ src/gui/widgets/CustomTabWidget.h
@@ -9,14 +9,14 @@ public:
     CustomTabWidget(QWidget* parent);
 
     //cpu usage in percentage, memoryUsage in megabytes
-    void setResourcesUsage(int cpuUsage, int memoryUsage);
+    void setResourcesUsage(double cpuUsage, int memoryUsage);
 protected:
     void paintEvent(QPaintEvent *event);
 private:
     static QColor RESOURCES_USAGE_BG_COLOR;
     static QColor RESOURCES_USAGE_TEXT_COLOR;
 
-    int cpuUsage;
+    double cpuUsage;
     int memoryUsage;
 };
